Baby step to resolving issue #1038
This does not change any #define's or logic other than cosmetically. It
replaces some hex constants with bit shifts to clarify the meaning of a
symbol. For example, #define NV_UNJUST 0x800000 with (1 << 23). It moves
some #define's to make the relationship with related symbols clearer. It
also adds more symbols to the list recognized by `nvflag_to_syms()`.

Related #1038

+++ src/cmd/ksh93/sh/nvdisc.c
@@ -798,10 +798,10 @@ void clone_all_disc(Namval_t *np, Namval_t *mp, int flags) {
 // NV_NODISC - discplines with funs non-zero will not be copied
 // NV_COMVAR - cloning a compound variable
 //
-int nv_clone(Namval_t *np, Namval_t *mp, int flags) {
+int nv_clone(Namval_t *np, Namval_t *mp, uint32_t flags) {
     Namfun_t *fp, *fpnext;
     const char *val = FETCH_VT(mp->nvalue, const_cp);
-    unsigned short flag = mp->nvflag;
+    uint32_t flag = mp->nvflag;
     size_t size = nv_size(mp);
 
     mp->nvshell = np->nvshell;
